+++ NopStarterKit.Web/Controllers/CheckoutController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using Nop.Integration.Umbraco.Core.Services;
+using Nop.Integration.Umbraco.Services.Order;
 using Umbraco.Web.Mvc;
 
 namespace NopStarterKit.Web.Controllers
@@ -11,16 +12,19 @@ namespace NopStarterKit.Web.Controllers
     public class CheckoutController : SurfaceController
     {
         private OrderService orderService;
-
+        private UserContext userContext;
         public CheckoutController()
         {
             this.orderService = new OrderService();
+            this.userContext = new UserContext();
         }
 
         // GET: Checkout
         public ActionResult ConfirmOrder()
         {
-            var id = orderService.CreateOrder();
+            int customerIdParseResult;
+            int.TryParse(userContext.CustomerId(), out customerIdParseResult);
+            var id = orderService.PlaceOrder(customerIdParseResult);
             return RedirectToAction("PayPalCreateRequest", "PayPal", new { orderId  = id });
         }
